pyproject.toml
[tool.poetry]
name = "mopsus-backend"
version = "0.1.0"
description = "\"mopsus backend project\""
authors = ["NicolasUrza <88108043+NicolasUrza@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
peewee = "^3.17.6"
google-cloud-bigquery = "^3.25.0"
pre-commit = "^3.8.0"

[tool.black]
line-length = 140
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "django"
combine_as_imports = true
include_trailing_comma = true
line_length = 120

[flake8]
max-line-length=140

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
pylint = "^3.2.7"
flake8 = "^7.1.1"
black = "^24.8.0"
pylint-django = "^2.5.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
